refactor: migrate RANSAC fitting tests to a dedicated module and streamline test structure
- Moved RANSAC fitting tests from `test_charge_sensor_compensation.py` to a new file `test_ransac_fitting.py` for better organization.
- Retained and updated existing tests to ensure they validate the functionality of `fit_compensation_gradient_ransac`, including gradient recovery and error handling.
- Improved overall clarity and maintainability of test cases by consolidating related tests into a dedicated module.
